% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-r0-of-time.R
\name{add_intervention}
\alias{add_intervention}
\title{Add an intervention}
\usage{
add_intervention(R, start, stop, c_reduce, r_reduce, delay, delay_dist = "exp")
}
\arguments{
\item{R}{Either a vector of the reproduction number for each step of the period being modeled
or a function that returns the reproduction number.}

\item{start}{The start time of the intervention}

\item{stop}{The end time of the intervention. For never enter Inf.}

\item{c_reduce}{The reduction in the proportion of daily contact as a result of
this measure}

\item{r_reduce}{An alternative to `pcontact_reduce`. The reduction in the reproduction
number.}

\item{delay}{The delay until full uptake of the measure.
For instant uptake (e.g. as a result of immediate strict enforcement by authorities)
enter 0, else some positive number that indicates the time to full uptake.}

\item{delay_dist}{The distribution of time to full uptake for non-zero delay times.
By default an exponential decay form "exp" is assumed where 99% uptake occurs by start+delay.
Alteratives include the uniform "unif" distribution where 100% uptake occurs at start+delay..}
}
\description{
Use `add_intervention` to model the impact of a epidemic control measure, for
example "social distancing", on the trajectory of a epidemic.
Is this measure adopted instantly or gradually over a number of units of time?
}
