% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-load-contact-matrix.R
\name{age_distributions_un}
\alias{age_distributions_un}
\title{List countries for which a (discrete) age distribution is available}
\usage{
age_distributions_un()
}
\description{
List countries for which a (discrete) age distribution is available
}
