% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{apply_intervention.default}
\alias{apply_intervention.default}
\title{Apply an intervention to a contact matrix or probability of transmission}
\usage{
\method{apply_intervention}{default}(x, reduce, dist)
}
\arguments{
\item{x}{contact matrix or probability of transmission}

\item{reduce}{amount by which we reduce the parameter under the intervention}

\item{dist}{distribution of groups in population}
}
\description{
Apply an intervention to a contact matrix or probability of transmission
}
