% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-load-contact-matrix.R
\name{available_contact_matrices}
\alias{available_contact_matrices}
\title{List countries for which a contact matrix is available}
\usage{
available_contact_matrices()
}
\description{
List countries for which a contact matrix is available
}
