% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{calculate_current_cm}
\alias{calculate_current_cm}
\title{Calculate current contact matrix}
\usage{
calculate_current_cm(cm, intervention, t, dist)
}
\arguments{
\item{cm}{...}

\item{intervention}{...}

\item{t}{...}
}
\description{
Calculate current contact matrix
}
