% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{calculate_current_pt}
\alias{calculate_current_pt}
\title{Calculate current probability of transmission}
\usage{
calculate_current_pt(pt, intervention, t)
}
\arguments{
\item{pt}{...}

\item{intervention}{...}

\item{t}{...}
}
\description{
Calculate current probability of transmission
}
