% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{calculate_reactive}
\alias{calculate_reactive}
\title{Calculate the effect of and update a reactive intervention at each time step}
\usage{
calculate_reactive(x, intervention, incRows, dist = NULL)
}
\arguments{
\item{x}{a contact matrix or the probability of transmission}

\item{intervention}{a reactive intervention}

\item{incRows}{the number of cases per group used to define the intervention
beginning and ending}

\item{dist}{population distribution of the groups}
}
\description{
Calculate the effect of and update a reactive intervention at each time step
}
