% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{calculate_reactive_pt}
\alias{calculate_reactive_pt}
\title{Calculate probability of transmission as a reactive function of incidence}
\usage{
calculate_reactive_pt(pt, intervention, incRows)
}
\arguments{
\item{pt}{...}

\item{intervention}{...}

\item{incRows}{...}
}
\description{
Calculate probability of transmission as a reactive function of incidence
}
