% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covis.R
\name{covis}
\alias{covis}
\title{Launch an interactive Shiny session}
\usage{
covis(example)
}
\arguments{
\item{example}{A Shiny app}
}
\description{
Launch an interactive Shiny session
}
\examples{
covis(dcm.models)
}
