% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coviz.R
\name{coviz}
\alias{coviz}
\title{Launch an interactive Shiny session}
\usage{
coviz(example)
}
\arguments{
\item{example}{A Shiny app}
}
\description{
Launch an interactive Shiny session
}
\examples{
coviz("dcm.models")
}
