% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-r0-of-time.R
\name{create_intervention}
\alias{create_intervention}
\title{Create an intervention}
\usage{
create_intervention(
  start,
  stop,
  c_reduce,
  start_delay,
  start_delay_dist = "logis",
  stop_delay,
  stop_delay_dist = "logis"
)
}
\arguments{
\item{start}{The start time of the intervention}

\item{stop}{The end time of the intervention.}

\item{c_reduce}{The relative reduction in the proportion of daily contact as a result of
this measure, e.g. enter 0.8 to indicate contact has been reduced to 80% of pre-intervention
levels}

\item{delay}{The delay until full uptake of the measure.
For instant uptake (e.g. as a result of immediate strict enforcement by authorities)
enter 0, else some positive number that indicates the time to full uptake.}

\item{delay_dist}{The distribution of time to full uptake for non-zero delay times.
By default an exponential decay form "exp" is assumed where 99% uptake occurs by start+delay.
Alteratives include the uniform "unif" distribution where 100% uptake occurs at start+delay..}
}
\description{
Use `create_intervention` to model the impact of a epidemic control measure, for
example "social distancing", on the trajectory of a epidemic.
Is this measure adopted instantly or gradually over a number of units of time?
}
\examples{

}
