% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot.R
\name{dcm2df}
\alias{dcm2df}
\title{Create ggplot2 ready dataframe: wrapper around tidyr's pivot_longer}
\usage{
dcm2df(mod)
}
\arguments{
\item{mod}{DCM model}
}
\description{
Create ggplot2 ready dataframe: wrapper around tidyr's pivot_longer
}
