% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-vaccine-model.R
\name{euler1}
\alias{euler1}
\title{Euler summation solving with intervention update}
\usage{
euler1(y, times, func, parms)
}
\arguments{
\item{y}{initial state}

\item{times}{time steps over which to sum}

\item{func}{difference equation, see `seir_cv_model`}

\item{parms}{parameters to pass to func, needs to be a list}
}
\description{
Euler summation solving with intervention update
}
