% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{exp_decay}
\alias{exp_decay}
\title{Exponential decay for interventions}
\usage{
exp_decay(delay, start_stop)
}
\arguments{
\item{delay}{delay until start of intervention}

\item{start_stop}{whether at beginning or end of intervention}
}
\description{
Exponential decay for interventions
}
