% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-load-contact-matrix.R
\name{import_age_distribution}
\alias{import_age_distribution}
\title{Load age distribution}
\usage{
import_age_distribution(country)
}
\arguments{
\item{country}{A character indicating which country.}
}
\description{
Load age distribution
}
\examples{
import_age_distribution("Australia")

}
