% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-load-contact-matrix.R
\name{import_total_population}
\alias{import_total_population}
\title{Load population estimates}
\usage{
import_total_population(country)
}
\arguments{
\item{country}{A character indicating which country.}
}
\description{
Load population estimates
}
\examples{
import_total_population("Australia")

}
