% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-r0-of-time.R
\name{logis_decay}
\alias{logis_decay}
\title{Logistic decay for interventions}
\usage{
logis_decay(delay, start_stop)
}
\arguments{
\item{delay}{delay until start of intervention}

\item{start_stop}{whether at beginning or end of intervention}
}
\description{
Logistic decay for interventions
}
