% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-vaccine-allocation.R
\name{mm}
\alias{mm}
\title{Michaelis–Menten like dynamics}
\usage{
mm(p, half)
}
\arguments{
\item{p}{proportion vaccinated}

\item{half}{point at which vaccination rate halfs}
}
\value{

}
\description{
Michaelis–Menten like dynamics
}
