% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot.R
\name{plot.contact_matrix}
\alias{plot.contact_matrix}
\title{Plot a contact matrix}
\usage{
\method{plot}{contact_matrix}(x, ...)
}
\arguments{
\item{x}{the contact matrix}
}
\description{
Plot a contact matrix
}
