% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seiqhrf-model.R
\name{seihrq_model}
\alias{seihrq_model}
\title{SEIHR-Q differential equations}
\usage{
seihrq_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E,I1,I2,H,Rh,R,Iq1,Iq2,Hq,Rqh,Rq)}

\item{param}{The model parameters (R0,beta,sigma,gamma1,gamma2,gamma3,Qeff,Heff,rho,alpha,eta)}
}
\value{
derivatives of SEIHR-Q model states with respect to time
}
\description{
SEIHR-Q differential equations
}
