% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seiqhrf-model.R
\name{seihrq_param}
\alias{seihrq_param}
\title{SEIHR-Q model parameters}
\usage{
seihrq_param(
  R0,
  beta,
  sigma,
  gamma1,
  gamma2,
  gamma3,
  Qeff,
  Heff,
  rho,
  alpha,
  eta
)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> E), can be a function of t.}

\item{beta}{Rate of potential new infections per infected (S -> E), can be a function of t.
Pass either R0 or beta.}

\item{sigma}{Inverse of the average length of latent period (E -> I1)}

\item{gamma1}{Inverse of the average length of first infectious period (I1 -> I2 or H)}

\item{gamma2}{Inverse of the average length of second infectious period (I2 -> R)}

\item{gamma3}{Inverse of the average length of second infectious period (for hospitalised) (H -> R or F)}

\item{Qeff}{Effect of quarantine on infectiousness, should be in (0,1)}

\item{Heff}{Effect of hospitalisation on infectiousness, should be in (0,1)}

\item{rho}{Proportion of people who enter quarantine after exposure, should be in (0,1) (E -> Iq1 or I1)}

\item{alpha}{Proportion of infected requiring hospitalisation, should be in (0,1) (I1 -> I2 or H)}

\item{eta}{Case fatality rate, should be in (0,1) (H -> F or Rh). Note F is calculated at each step, there is no dF.}
}
\value{
List of SEIHR-Q model parameters
}
\description{
Setup function
}
