% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-example-parameters.R
\name{seihrq_param_set}
\alias{seihrq_param_set}
\title{SEIHR-Q model preset parameter sets}
\usage{
seihrq_param_set(set)
}
\arguments{
\item{set}{the parameter set to use}
}
\description{
SEIHR-Q model preset parameter sets
}
