% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{seimrqc_model}
\alias{seimrqc_model}
\title{SEIMR-QC differential equations}
\usage{
seimrqc_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E1,E2,I1,I2,R,M,Rm,Eq1,Eq2,Iq1,Iq2,Rq,Mq,Rqm)}

\item{param}{The model parameters (R0,beta,sigma,gamma1,gamma2,gamma3,Qeff,Heff,rho,alpha,eta)}
}
\value{
derivatives of SEIMR-QC model states with respect to time
}
\description{
SEIMR-QC differential equations
}
