% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{seimrqc_param}
\alias{seimrqc_param}
\title{SEIMR-QC model parameters}
\usage{
seimrqc_param(
  R0,
  lambdaimp,
  sigma1,
  sigma2,
  gamma1,
  gamma2,
  Qeff,
  Meff,
  rho,
  alpha,
  eta,
  delta,
  kappa,
  pm
)
}
\arguments{
\item{R0}{Basic reproduction number (S -> E)}

\item{lambdaimp}{The force of infection from importated cases.}

\item{sigma1}{Inverse of the average length of first latent period (E1 -> E2)}

\item{sigma2}{Inverse of the average length of second latent period (E2 -> I1)}

\item{gamma1}{Inverse of the average length of first infectious period (I1 -> I2 or H)}

\item{gamma2}{Inverse of the average length of second infectious period (I2 -> R)}

\item{Qeff}{Effect of quarantine on infectiousness, should be in (0,1)}

\item{Meff}{Effect of medical management on infectiousness, should be in (0,1)}

\item{rho}{the proportion of contacts (of ascertained cases) that will self-quarantine (0,1)}

\item{alpha}{Proportion of infected requesting medical assistance, should be in (0,1)}

\item{eta}{...}

\item{delta}{The duration of quarantine for contacts of ascertained cases (14 days)}

\item{kappa}{The per-person contact rate (20 people per day)}

\item{pm}{The probability of presenting cases being effectively managed}
}
\value{
List of seimrqc model parameters
}
\description{
Setup function
}
