% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{seimrqc_state0}
\alias{seimrqc_state0}
\title{SEIMR-QC model inital state}
\usage{
seimrqc_state0(
  S0,
  E0,
  I10 = 0,
  I20 = 0,
  H0 = 0,
  Rh0 = 0,
  R0 = 0,
  Iq10 = 0,
  Iq20 = 0,
  Hq0 = 0,
  Rqh0 = 0,
  Rq0 = 0
)
}
\arguments{
\item{S0}{Initial number of susceptibles}

\item{E0}{Initial number of exposed}

\item{I10}{Initial number of infected (period 1)}

\item{I20}{Initial number of infected (period 2)}

\item{H0}{Initial number of hospitalised}

\item{Rh0}{Initial number of recovered hospitalisations}

\item{R0}{Initial number of recovered}

\item{Iq10}{Initial number of infected (period 1) (who had self quarantined)}

\item{Iq20}{Initial number of infected (period 2) (who had self quarantined)}

\item{Hq0}{Initial number of hospitalised (who had self quarantined)}

\item{Rqh0}{Initial number of recovered hospitalisations (who had self quarantined)}

\item{Rq0}{Initial number of recovered (who had self quarantined)}
}
\value{
List of seimrqc model initial states
}
\description{
Setup function
}
