% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-example-parameters.R
\name{seiqhrf_param_set}
\alias{seiqhrf_param_set}
\title{SEIQHRF model preset parameter sets}
\usage{
seiqhrf_param_set(set)
}
\arguments{
\item{set}{the parameter set to use}
}
\description{
SEIQHRF model preset parameter sets
}
