% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir1-model.R
\name{seir1_model}
\alias{seir1_model}
\title{SEIR model with quarantine and hospitalisation (SEIR-QH) differential equations}
\usage{
seir1_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E1,E2,I1,I2,H,Rh,R,Eq1,Eq2,Iq1,Iq2,Hq,Rqh,Rq)}

\item{param}{The model parameters (R0,beta,sigma1,sigma2,gamma1,gamma2,gamma3,Qeff,Heff,rho,alpha,eta)}
}
\value{
derivatives of SEIR-QH model states with respect to time
}
\description{
SEIR model with quarantine and hospitalisation (SEIR-QH) differential equations
}
