% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir1-model.R
\name{seir1_param}
\alias{seir1_param}
\title{SEIR model with quarantine and hospitalisation (SEIR-QH) parameters}
\usage{
seir1_param(
  R0,
  beta,
  sigma1,
  sigma2,
  gamma1,
  gamma2,
  gamma3,
  Qeff,
  Heff,
  rho,
  alpha,
  eta
)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> E1), can be a function of t.}

\item{beta}{Rate of potential new infections per infected (S -> E1), can be a function of t.
Pass either R0 or beta.}

\item{sigma1}{Inverse of the average length of first latent period (E1 -> E2)}

\item{sigma2}{Inverse of the average length of second latent period (E2 -> I1)}

\item{gamma1}{Inverse of the average length of first infectious period (I1 -> I2)}

\item{gamma2}{Inverse of the average length of second infectious period (I2 -> R)}

\item{gamma3}{Inverse of the average length of second infectious period for hospitalised cases (H -> R/F)}

\item{Qeff}{Effect of quarantine on infectiousness, should be in (0,1)}

\item{Heff}{Effect of hospitalisation on infectiousness, should be in (0,1)}

\item{rho}{Proportion of people who enter quarantine after exposure, should be in (0,1) (E -> Iq1 or I1)}

\item{alpha}{Proportion of infected requiring hospitalisation, should be in (0,1) (I1 -> I2 or H)}

\item{eta}{Case fatality rate, should be in (0,1) (H -> F or Rh). Note F is calculated at each step, there is no dF.}
}
\value{
List of SEIHR-Q model parameters
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Setup function
}
