% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir1-model.R
\name{seir1_state0}
\alias{seir1_state0}
\title{SEIR model with quarantine and hospitalisation (SEIR-QH) inital state}
\usage{
seir1_state0(
  S,
  E1,
  E2,
  I1 = 0,
  I2 = 0,
  H = 0,
  Rh = 0,
  R = 0,
  Eq1 = 0,
  Eq2 = 0,
  Iq1 = 0,
  Iq2 = 0,
  Hq = 0,
  Rqh = 0,
  Rq = 0
)
}
\arguments{
\item{S}{Initial number of susceptibles}

\item{E1}{Initial number of exposed (period 1)}

\item{E2}{Initial number of exposed (period 2)}

\item{I1}{Initial number of infected (period 1)}

\item{I2}{Initial number of infected (period 2)}

\item{H}{Initial number of hospitalised}

\item{Rh}{Initial number of recovered hospitalisations}

\item{R}{Initial number of recovered}

\item{Eq1}{Initial number of exposed (period 1)}

\item{Eq2}{Initial number of exposed (period 2)}

\item{Iq1}{Initial number of infected (period 1) (who were in isolation)}

\item{Iq2}{Initial number of infected (period 2) (who were in isolation)}

\item{Hq}{Initial number of hospitalised (who were in isolation)}

\item{Rqh}{Initial number of recovered hospitalisations (who were in isolation)}

\item{Rq}{Initial number of recovered (who were in isolation)}
}
\value{
List of SEIR 1 model initial states
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Setup function
}
