% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R, R/1-seir2-model.R
\name{seir2_model}
\alias{seir2_model}
\title{Doherty differential equations}
\usage{
seir2_model(t, state_t0, param)

seir2_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E,I,R)}

\item{param}{The model parameters (R0,beta,sigma,gamma)}
}
\value{
derivatives of Doherty model states with respect to time

derivatives of SEIR model states with respect to time
}
\description{
Doherty differential equations

SEIR differential equations
}
