% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{seir2_state0}
\alias{seir2_state0}
\title{Doherty model initial state}
\usage{
seir2_state0(
  S,
  E1,
  E2 = 0,
  I1 = 0,
  I2 = 0,
  R = 0,
  M = 0,
  Rm = 0,
  Eq1 = 0,
  Eq2 = 0,
  Iq1 = 0,
  Iq2 = 0,
  Rq = 0,
  Mq = 0,
  Rqm = 0,
  CTm = 2,
  CTnm = 20
)
}
\arguments{
\item{S}{Initial number of susceptibles}

\item{E1}{Initial number of exposed (latent period 1)}

\item{E2}{Initial number of exposed (latent period 2)}

\item{I1}{Initial number of infected (period 1)}

\item{I2}{Initial number of infected (period 2)}

\item{R}{Initial number of removed}

\item{M}{Initial number of under medical management}

\item{Rm}{Initial number of recovered who sought medical treatment}

\item{Eq1}{Initial number of exposed (latent period 1) who self quarantined after contact}

\item{Eq2}{Initial number of exposed (latent period 2) who self quarantined after contact}

\item{Iq1}{Initial number of infected (period 1) who self quarantined after contact}

\item{Iq2}{Initial number of infected (period 2) who self quarantined after contact}

\item{Rq}{Initial number of removed who self quarantined after contact}

\item{Mq}{Initial number of under medical management who self quarantined after contact}

\item{Rqm}{Initial number of recovered who sought medical treatment and self quarantined after contact}
}
\value{
List of seir2 model initial states
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Setup function
}
