% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-model.R
\name{seir_c_model}
\alias{seir_c_model}
\title{SEIR with heterogeneous contact differential equations}
\usage{
seir_c_model(t, y, parms)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E,I,R)}

\item{param}{The model parameters (R0,sigma,gamma,cm,dist,intervention)}
}
\value{
derivatives of SEIR model states with respect to time
}
\description{
SEIR with heterogeneous contact differential equations
}
