% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-model.R
\name{seir_c_param}
\alias{seir_c_param}
\title{Age structured SEIR model parameters}
\usage{
seir_c_param(R0, sigma, gamma, cm, dist, intervention = NULL)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> E).}

\item{sigma}{Inverse of the average length of the incubation period (E -> I)}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}

\item{cm}{Contact matrix or a named list of contact matrices where the sum is taken to
be all contacts in the population.}

\item{dist}{Proportion of the population in each age category. Should sum to 1.}

\item{intervention}{An intervention created using `create_intervention` or named list
of interventions with the names matching the list of contact matrices.}
}
\value{
List of SEIR model parameters
}
\description{
Setup function
}
