% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-vaccine-model.R
\name{seir_cv_param}
\alias{seir_cv_param}
\title{Age structured SEIR model parameters}
\usage{
seir_cv_param(
  R0,
  sigma,
  gamma,
  cm,
  dist,
  vaceff1,
  vaceff2,
  vaceff3,
  nvac,
  vac_alloc,
  n_imp,
  contact_intervention = NULL,
  transmission_intervention = NULL,
  im = NULL
)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> E).}

\item{sigma}{Inverse of the average length of the latent period (E -> I)}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}

\item{cm}{Contact matrix or a named list of contact matrices where the sum is taken to
be all contacts in the population.}

\item{dist}{Proportion of the population in each age category. Should sum to 1.}

\item{vaceff1}{Vaccine effectiveness: P(infection) of vaccinated (a vector of length J)}

\item{vaceff2}{Vaccine effectiveness: S -> Sv (a vector of length J)}

\item{vaceff3}{Vaccine effectiveness: Iv -> S/Sv (a vector of length J)}

\item{nvac}{Number vaccinated.}

\item{n_imp}{Number of infection imported cases (a function of time).}

\item{contact_intervention}{An intervention created using `contact_intervention` or named list
of interventions with the names matching the list of contact matrices.}

\item{transmission_intervention}{An intervention created using `create_intervention` or named list
of interventions with the names matching the list of contact matrices.}

\item{im}{Relative infectiousness matrix. Default is 1.}
}
\value{
List of SEIR model parameters
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

Define parameter values for age structured SEIR models
}
