% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-vaccine-model.R
\name{seir_cv_state0}
\alias{seir_cv_state0}
\title{Age structured SEIR model inital state}
\usage{
seir_cv_state0(S, E, I, R, Sv, Ev, Iv, Rv)
}
\arguments{
\item{S}{Initial number of susceptibles}

\item{E}{Initial number of exposed (latent period)}

\item{I}{Initial number of infectious}

\item{R}{Initial number of removed}
}
\value{
List of SEIR model initial states
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

Define intial state values for age structured SEIR models
}
