% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-model.R
\name{seir_model}
\alias{seir_model}
\title{SEIR differential equations}
\usage{
seir_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E,I,R)}

\item{param}{The model parameters (R0,beta,sigma,gamma)}
}
\value{
derivatives of SEIR model states with respect to time
}
\description{
SEIR differential equations
}
