% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-model.R
\name{seir_param}
\alias{seir_param}
\title{SEIR model parameters}
\usage{
seir_param(R0, beta, sigma, gamma)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> E), can be a function of t.}

\item{beta}{Rate of potential new infections per infected (S -> E), can be a function of t.
Pass either R0 or beta.}

\item{sigma}{Inverse of the average length of latent period (E -> I)}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}
}
\value{
List of SEIR model parameters
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Setup function
}
