% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seithr-qcm-model.R
\name{seithr_qc_model}
\alias{seithr_qc_model}
\title{SEITHR-QC with heterogeneous contact differential equations}
\usage{
seithr_qc_model(t, y, parms)
}
\arguments{
\item{t}{current time (dt scale)}

\item{state_t0}{Initial state (S,I,R)}

\item{param}{The model parameters (beta,R0,gamma,cm)}
}
\value{
derivatives of SEIR model states with respect to time
}
\description{
SEITHR-QC with heterogeneous contact differential equations
}
