% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seithr-qcm-model.R
\name{seithr_qc_state0}
\alias{seithr_qc_state0}
\title{SEIR-QHT model inital state}
\usage{
seithr_qc_state0(
  S,
  E1,
  E2,
  I1m,
  I1s,
  I2,
  I2i,
  H,
  Sc,
  Ec1,
  Ec2,
  Ic1m,
  Ic1s,
  Ic2,
  Hc,
  R,
  Rh,
  Tested,
  Detect,
  NotDetect
)
}
\arguments{
\item{Rh}{}
}
\value{
List of SEIR-QHT model initial states
}
\description{
Setup function
}
