% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seiqhrf-model.R
\name{simulate_seihrq}
\alias{simulate_seihrq}
\title{Simulate a deterministic SEIHR-Q model}
\usage{
simulate_seihrq(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see seihrq_state0)}

\item{param}{Model parameters (see seihrq_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
\itemize{
\item S = Susceptible
\item E = Exposed (Asymptomatic but infectious)
\item I1/Iq1 = First infectious period
\item I2/Iq2 = Second infection period
\item H/Hq = Hospitalised
\item R/Rq/Rh/Rqh = Removed
\item F = Case fatality (derived number, not a modelled compartment)
}
}
\details{
Adapted version of the stochastic individual contact seihrq model outlined in Churches & Jorm (2020).
}
\examples{


}
