% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{simulate_seimrqc}
\alias{simulate_seimrqc}
\title{Simulate a deterministic SEIMR-QC ("Doherty" or "Moss") model}
\usage{
simulate_seimrqc(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see seimrqc_state0)}

\item{param}{Model parameters (see seimrqc_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
\itemize{
\item S = Susceptible
\item E1/Eq1 = Exposed (asymptomatic; not infectious)
\item E2/Eq2 = Exposed (asymptomatic; infectious)
\item I1/Iq1 = First infectious period
\item I2/Iq2 = Second infection period
\item M/Mq = Under medical management (hospitalised; GP)
\item R/Rq/Rm/Rqm = Recovered
}
}
\details{
Implementation of the model outlined in Moss et al (2020).
}
\examples{
param <- seimrqc_param(R0=2.5,sigma=0.3,gamma1=0.3,gamma2=0.3,gamma3=0.3,Qeff=0.5,Heff=0.99,rho=0.1,alpha=0.2,eta=0.01)
state0 <- seimrqc_state0(S0=100,E0=1)
res <- simulate_seimrqc(t = 100,state_t0 = state0,param = param)
plot(res,c("S","E","Is","Hp","Rc","F"))

}
