% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-model.R
\name{simulate_seir}
\alias{simulate_seir}
\title{Simulate a deterministic SEIR model}
\usage{
simulate_seir(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see seir_state0)}

\item{param}{Model parameters (see seir_param)}
}
\value{
Object of class covoid and dcm (from the package EpiModels)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

S = Susceptible
E = Exposed (non-infectious)
I = Infectious
R = Removed
}
\examples{
param <- seir_param(R0 = 2.5,gamma = 0.1,sigma=0.1)
state0 <- seir_state0(S = 100,E = 1, I = 0,R = 0)
res <- simulate_seir(t = 100,state_t0 = state0,param = param)
plot(res,c("S","E","I","R"))

# if the reproduction number is time varying (e.g. due to control measures)
R_t <- function(t) {
if (t < 30) 2.5
else 1.1
}
param <- seir_param(R0 = R_t,gamma = 0.1,sigma=0.1)
state0 <- seir_state0(S = 100,E = 1, I = 0,R = 0)
res <- simulate_seir(t = 100,state_t0 = state0,param = param)
plot(res,c("S","E","I","R"))

}
