% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir1-model.R
\name{simulate_seir1}
\alias{simulate_seir1}
\title{Simulate a deterministic SEIR model with quarantine and hospitalisation (SEIR-QH)}
\usage{
simulate_seir1(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see ?seir1_state0)}

\item{param}{Model parameters (see ?seir1_param)}
}
\value{
Object of class covoid
}
\description{
\itemize{
\item S = Susceptible
\item Eq1/Eq1 = First latent period (asymptomatic; uninfectious)
\item E2/Eq2 = Second latent period (asymptomatic; infectious)
\item I2/Iq2 = Second infectious period (symptomatic; infectious)
\item I2/Iq2 = Third infection period (symptomatic; infectious)
\item H/Hq = Hospitalised
\item R/Rq/Rh/Rqh = Removed
\item F = Case fatality (derived number, not a modelled compartment)
}
}
\details{
Adapted version of the stochastic individual contact seir1 model outlined in Churches & Jorm (2020).
}
\examples{

}
