% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-model.R
\name{simulate_seir_c}
\alias{simulate_seir_c}
\title{Simulate a deterministic age structured SEIR model}
\usage{
simulate_seir_c(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see ?seir_state0)}

\item{param}{Model parameters (see ?seir_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
S = Susceptible
E = Exposed (incubating)
I = Infectious
R = Removed (recovered or died)
}
\examples{

cm_oz <- import_contact_matrix("Australia","general")
dist_oz <- import_age_distribution("Australia")
param <- seir_c_param(R0 = 2.5,sigma=0.1,gamma = 0.1,cm=cm_oz,dist=dist_oz)
nJ = ncol(cm_oz)
S = rep(100,nJ)
E = rep(1,nJ)
I = rep(1,nJ)
R = rep(0,nJ)
state0 <- seir_c_state0(S = S,E = E,I = I,R = R)
res <- simulate_seir_c(t = 150,state_t0 = state0,param = param)
plot(res,y=c("S","E","I","R"),main="Heterogeneous mixing")

}
