% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-model.R
\name{simulate_sir}
\alias{simulate_sir}
\title{Simulate a deterministic SIR model}
\usage{
simulate_sir(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see sir_state0)}

\item{param}{Model parameters (see sir_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
S = Susceptible
I = Infectious
R = Removed
}
