% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-cm-model.R
\name{simulate_sir_c}
\alias{simulate_sir_c}
\title{Simulate a deterministic age structured SIR model}
\usage{
simulate_sir_c(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see ?sir_state0)}

\item{param}{Model parameters (see ?sir_param)}
}
\value{
Object of class covoid and dcm (from the package EpiModels)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

S = Susceptible
I = Infectious
R = Removed
}
\examples{

cm_oz <- import_contact_matrix("Australia","general")
nJ <- ncol(cm_oz)
dist_oz <- import_age_distribution("Australia")
param <- sir_c_param(R0 = 2.5,gamma = 0.1,cm=cm_oz,dist=dist_oz)
nJ = ncol(cm_oz)
S = rep(100,nJ)
I = rep(1,nJ)
R = rep(0,nJ)
state0 <- sir_c_state0(S = S,I = I,R = R)
res <- simulate_sir_c(t = 150,state_t0 = state0,param = param)
plot(res,y=c("S","I","R"),main="Heterogeneous mixing")

# compare with homogeneous
param <- sir_param(R0 = 2.5,gamma = 0.1)
state0 <- sir_state0(S = 1600,I = 16,R = 0)
res <- simulate_sir(t = 150,state_t0 = state0,param = param)
plot(res,c("S","I","R"),main="Homogeneous mixing")

}
