% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-cm-model.R
\name{sir_c_model}
\alias{sir_c_model}
\title{SIR with heterogeneous contact differential equations}
\usage{
sir_c_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,I,R)}

\item{param}{The model parameters (beta,R0,gamma,cm)}
}
\value{
derivatives of SIR model states with respect to time
}
\description{
SIR with heterogeneous contact differential equations
}
