% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-cm-model.R
\name{sir_c_param}
\alias{sir_c_param}
\title{Age structured SIR model parameters}
\usage{
sir_c_param(
  R0,
  gamma,
  cm,
  dist,
  contact_intervention = NULL,
  transmission_intervention = NULL,
  im = NULL
)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> I), can be a function of t.}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}

\item{cm}{Contact matrix or a named list of contact matrices where the sum is taken to
be all contacts in the population.}

\item{dist}{Proportion of the population in each age category. Should sum to 1.}

\item{contact_intervention}{An intervention created using `contact_intervention` or named list
of interventions with the names matching the list of contact matrices.}

\item{transmission_intervention}{An intervention created using `create_intervention` or named list
of interventions with the names matching the list of contact matrices.}

\item{im}{Relative infectiousness matrix. Default is 1.}
}
\value{
List of SIR model parameters
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

Define parameter values for age structured SIR models
}
