% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-model.R
\name{sir_model}
\alias{sir_model}
\title{SIR differential equations}
\usage{
sir_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,I,R)}

\item{param}{The model parameters (beta,R0,gamma)}
}
\value{
derivatives of SIR model states with respect to time
}
\description{
SIR differential equations
}
