% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-model.R
\name{sir_param}
\alias{sir_param}
\title{SIR model parameters}
\usage{
sir_param(R0, beta, gamma)
}
\arguments{
\item{R0}{Basic reproduction number (S -> I)}

\item{beta}{Rate of potential new infections per infected (S -> I)}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}
}
\value{
List of SIR model parameters
}
\description{
Setup function
}
