% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-model.R
\name{sir_state0}
\alias{sir_state0}
\title{SIR model inital state}
\usage{
sir_state0(S, I, R = 0)
}
\arguments{
\item{S}{Initial number of susceptibles}

\item{I}{Initial number of infected}

\item{R}{Initial number of removed}
}
\value{
List of SIR model initial states
}
\description{
Setup function
}
