% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{transmission_intervention}
\alias{transmission_intervention}
\title{Transmission intervention}
\usage{
transmission_intervention(
  start,
  stop,
  reduce,
  start_delay = 0,
  start_delay_dist = "logis",
  stop_delay = 0,
  stop_delay_dist = "logis"
)
}
\arguments{
\item{start}{The start time of the intervention}

\item{stop}{The end time of the intervention.}

\item{reduce}{The relative reduction in the probability of transmission
as a result of this measure, e.g. enter 0.8 to indicate contact has been reduced to 80% of pre-intervention
levels}

\item{delay}{The delay until full uptake of the measure.
For instant uptake (e.g. as a result of immediate strict enforcement by authorities)
enter 0, else some positive number that indicates the time to full uptake.}

\item{delay_dist}{The distribution of time to full uptake for non-zero delay times.}
}
\description{
Use `transmission_intervention` to model the impact of a epidemic control measure, for
example wearing of PPE, on the trajectory of a epidemic.
Is this measure adopted instantly or gradually over a number of units of time?
}
\examples{

}
