% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-vaccine-allocation.R
\name{vaccination_allocation_mm}
\alias{vaccination_allocation_mm}
\title{Vaccine allocation with Michaelis–Menten tails}
\usage{
vaccination_allocation_mm(t, n, s, vac_params)
}
\arguments{
\item{t}{time t}

\item{n}{number of available vaccines at time t}

\item{s}{number of people in group j}

\item{vac_params}{a list including:
p: probability of group j getting vaccinates
s0: group size at time 0
half: percentage of population at which slow down begins}
}
\value{

}
\description{
Vaccine allocation with Michaelis–Menten tails
}
