% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-vaccine-allocation.R
\name{vaccination_allocation_pp}
\alias{vaccination_allocation_pp}
\title{Vaccine allocataion incorporating phases and propensity to be vaccinated}
\usage{
vaccination_allocation_pp(t, n, s, vac_params)
}
\arguments{
\item{t}{Time t}

\item{n}{Number of vaccine doses available at time t}

\item{s}{A vector with the number of people in group j on day t}

\item{vac_params}{A list specifying openDay (Day the vaccines open for each phase) and propensity (The probability of being available for vaccination on a single day)}
}
\value{
A vector distributing the \code{n} vaccine doses across the \code{J} groups
}
\description{
Distribute vaccines across different phases with time-dependent eligibility for different phases and varying vaccine willingness within phases
}
\examples{
# Distribute 20 vaccines across two phases with a high and low propensity subgroup in each phase
vax_allocation(t=1, n=20, s=rep(100,4), vac_params = list(openDay = c(1,8), propensity = c(.3,.7)))
}
