% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{add_intervention}
\alias{add_intervention}
\title{Add an intervention (deprecated)}
\usage{
add_intervention(
  R,
  start,
  stop,
  c_reduce,
  r_reduce,
  start_delay,
  start_delay_dist = "logis",
  stop_delay,
  stop_delay_dist = "logis"
)
}
\arguments{
\item{R}{Either a vector of the (effective) reproduction number for each time step of the period being modeled
or a function that returns the reproduction number.}

\item{start}{The start time of the intervention}

\item{stop}{The end time of the intervention.}

\item{c_reduce}{The reduction in the proportion of daily contact as a result of
this measure}

\item{r_reduce}{An alternative to `pcontact_reduce`. Not implemented!
The expected reduction in the effective reproduction number.}

\item{delay}{The delay until full uptake of the measure.
For instant uptake (e.g. as a result of immediate strict enforcement by authorities)
enter 0, else some positive number that indicates the time to full uptake.}

\item{delay_dist}{The distribution of time to full uptake for non-zero delay times.
By default an exponential decay form "exp" is assumed where 99% uptake occurs by start+delay.
Alteratives include the uniform "unif" distribution where 100% uptake occurs at start+delay..}
}
\description{
\lifecycle{deprecated}

Use `add_intervention` to model the impact of a epidemic control measure, for
example "social distancing", on the trajectory of a epidemic.
Is this measure adopted instantly or gradually over a number of units of time?
}
\examples{
# function as input
R <- function(t) 2.5
R_updated <- add_intervention(R,5,25,0.5,start_delay=10,stop_delay=10)
R_updated(9)
plot(1:50,sapply(1:50,R_updated))  # the function is not vectorised

# vector as input
R <- rep(2.5,30)
R_updated <- add_intervention(R,5,25,0.5,delay=10)
R_updated[9]
plot(1:30,R_updated)

}
