% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-load-contact-matrix.R
\name{import_contact_matrix}
\alias{import_contact_matrix}
\title{Import a contact matrix for country and setting}
\usage{
import_contact_matrix(country, setting)
}
\arguments{
\item{country}{A character indicating which country.}

\item{setting}{The settting for the desired contact matrix.
Can be one of "general","work","school","home","other"}
}
\description{
Each row i of a column j of the contact matrix corresponds to the number of contacts
made by an individual in group j with an individual in group i.
}
\examples{
import_contact_matrix("Australia","general")

}
