% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{reactive_intervention}
\alias{reactive_intervention}
\title{Reactive intervention}
\usage{
reactive_intervention(threshold, reduce, state)
}
\arguments{
\item{threshold}{threshold (I - prevalence) for an intervention to commence}

\item{reduce}{reduction in contact or transmission probability}

\item{state}{state of the intervention: see `reactive_state`}
}
\description{
Reactive intervention
}
