% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-interventions.R
\name{reactive_intervention}
\alias{reactive_intervention}
\title{Reactive intervention}
\usage{
reactive_intervention(threshold, reduce)
}
\arguments{
\item{threshold}{incidence threshold for an intervention}

\item{reduce}{reduction in contact or transmission probability}
}
\description{
Reactive intervention
}
