% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{seimrqc_state0}
\alias{seimrqc_state0}
\title{SEIMR-QC model inital state}
\usage{
seimrqc_state0(
  S0,
  E10,
  E20 = 0,
  I10 = 0,
  I20 = 0,
  M0 = 0,
  Rm0 = 0,
  R0 = 0,
  Eq10 = 0,
  Eq20 = 0,
  Iq10 = 0,
  Iq20 = 0,
  Mq0 = 0,
  Rqm0 = 0,
  Rq0 = 0,
  CTm0 = 0
)
}
\arguments{
\item{S0}{Initial number of susceptibles}

\item{E10}{Initial number of exposed (latent period 1)}

\item{E20}{Initial number of exposed (latent period 2)}

\item{I10}{Initial number of infected (period 1)}

\item{I20}{Initial number of infected (period 2)}

\item{M0}{Initial number of under medical management}

\item{Rm0}{Initial number of recovered who sought medical treatment}

\item{R0}{Initial number of removed}

\item{Eq10}{Initial number of exposed (latent period 1) who self quarantined after contact}

\item{Eq20}{Initial number of exposed (latent period 2) who self quarantined after contact}

\item{Iq10}{Initial number of infected (period 1) who self quarantined after contact}

\item{Iq20}{Initial number of infected (period 2) who self quarantined after contact}

\item{Mq0}{Initial number of under medical management who self quarantined after contact}

\item{Rqm0}{Initial number of recovered who sought medical treatment and self quarantined after contact}

\item{Rq0}{Initial number of removed who self quarantined after contact}
}
\value{
List of seimrqc model initial states
}
\description{
Setup function
}
