% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-doherty-model.R
\name{seir2_model}
\alias{seir2_model}
\title{Doherty differential equations}
\usage{
seir2_model(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state (S,E1,E2,I1,I2,R,M,Rm,Eq1,Eq2,Iq1,Iq2,Rq,Mq,Rqm)}

\item{param}{The model parameters (R0,sigma1,sigma2,gamma1,gamma2,gammaq1,gammaq2,Qeff,Heff,rho,alpha)}
}
\value{
derivatives of Doherty model states with respect to time
}
\description{
Doherty differential equations
}
