% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-model.R
\name{seir_c_state0}
\alias{seir_c_state0}
\title{SEIR model inital state}
\usage{
seir_c_state0(S, E, I, R)
}
\arguments{
\item{S}{Initial number of susceptibles}

\item{E}{Initial number of incubating}

\item{I}{Initial number of infectious}

\item{R}{Initial number of removed}
}
\value{
List of SEIR model initial states
}
\description{
Setup function
}
