% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-model.R
\name{seir_state0}
\alias{seir_state0}
\title{SEIR model inital state}
\usage{
seir_state0(S0, E0, I0 = 0, R0 = 0)
}
\arguments{
\item{S0}{Initial number of susceptibles}

\item{E0}{Initial number of exposed}

\item{I0}{Initial number of infected}

\item{R0}{Initial number of removed}
}
\value{
List of SEIR model initial states
}
\description{
Setup function
}
