% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seithr-qcm-model.R
\name{seithr_qc_param}
\alias{seithr_qc_param}
\title{SEIR-QHT model parameters}
\usage{
seithr_qc_param(
  R0,
  Qeff,
  Teff,
  Heff,
  delta,
  sigma1,
  sigma2,
  p_severe,
  p_isolate,
  gamma1,
  gamma2,
  cm,
  dist,
  contact_intervention = NULL,
  transmission_intervention = NULL,
  im = NULL
)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> I), can be a function of t.}

\item{Qeff}{...}

\item{Teff}{...}

\item{Heff}{...}

\item{delta}{...}

\item{sigma1}{...}

\item{sigma2}{...}

\item{p_severe}{...}

\item{p_isolate}{...}

\item{gamma1}{...}

\item{gamma2}{...}

\item{cm}{Contact matrix or a named list of contact matrices where the sum is taken to
be all contacts in the population.}

\item{dist}{Proportion of the population in each age category. Should sum to 1.}

\item{contact_intervention}{An intervention created using `contact_intervention` or named list
of interventions with the names matching the list of contact matrices.}

\item{transmission_intervention}{An intervention created using `create_intervention` or named list
of interventions with the names matching the list of contact matrices.}

\item{im}{Relative infectiousness matrix. Default is 1.}
}
\value{
List of SIR model parameters
}
\description{
Setup function
}
