% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seiqhrf-model.R
\name{simulate_seihrq}
\alias{simulate_seihrq}
\title{Simulate a deterministic SEIHR-Q model}
\usage{
simulate_seihrq(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see seihrq_state0)}

\item{param}{Model parameters (see seihrq_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
\itemize{
\item S = Susceptible
\item E = Exposed (Asymptomatic but infectious)
\item I1/Iq1 = First infectious period
\item I2/Iq2 = Second infection period
\item H/Hq = Hospitalised
\item R/Rq/Rh/Rqh = Removed
\item F = Case fatality (derived number, not a modelled compartment)
}
}
\details{
Adapted version of the stochastic individual contact seihrq model outlined in Churches & Jorm (2020).
}
\examples{
param <- seihrq_param(R0=2.5,sigma=0.3,gamma1=0.3,gamma2=0.3,gamma3=0.3,Qeff=0.5,Heff=0.99,rho=0.1,alpha=0.2,eta=0.01)
state0 <- seihrq_state0(S0=100,E0=1)
res <- simulate_seihrq(t = 100,state_t0 = state0,param = param)
plot(res,c("S","E","Is","Hp","Rc","F"))

}
