% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-model.R
\name{simulate_seir}
\alias{simulate_seir}
\title{Simulate a deterministic SEIR model}
\usage{
simulate_seir(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see seir_state0)}

\item{param}{Model parameters (see seir_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
S = Susceptible
E = Exposed (not infectious)
I = Infectious
R = Removed
}
\examples{
param <- seir_param(R0 = 2.5,gamma = 0.1,sigma=10)
state0 <- seir_state0(S0 = 100,E0 = 1, I0 = 0,R0 = 0)
res <- simulate_seir(t = 100,state_t0 = state0,param = param)

}
