% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir1-model.R
\name{simulate_seir1}
\alias{simulate_seir1}
\title{Simulate a deterministic SEIR model with quarantine and hospitalisation (SEIR-QH)}
\usage{
simulate_seir1(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see ?seir1_state0)}

\item{param}{Model parameters (see ?seir1_param)}
}
\value{
Object of class covoid
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

\itemize{
\item S = Susceptible
\item Eq1/Eq1 = First latent period (asymptomatic; uninfectious)
\item E2/Eq2 = Second latent period (asymptomatic; infectious)
\item I2/Iq2 = Second infectious period (symptomatic; infectious)
\item I2/Iq2 = Third infection period (symptomatic; infectious)
\item H/Hq = Hospitalised
\item R/Rq/Rh/Rqh = Removed
\item F = Case fatality (derived number, not a modelled compartment)
}

Adapted version of the stochastic individual contact seir1 model outlined in Churches & Jorm (2020).
}
\examples{
state0 <- seir1_state0(S = 1e5, E1 = 90, E2 = 40)
param <- seir1_param(R0=2.5,sigma1=0.2,sigma2=0.2,gamma1=0.2,gamma2=0.2,gamma3=0.2,
                     Qeff=0.5,Heff=0.9,rho=0.1,alpha=0.1,eta=0.02)
res <- simulate_seir1(t = 250,state_t0 = state0,param = param)
plot(res,y=c("S","E","I","Recov","Fatal"),main="Expanded SEIR model I")

}
