% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-seir-cm-vaccine-model.R
\name{simulate_seir_cv}
\alias{simulate_seir_cv}
\title{Simulate a deterministic age structured SEIR model}
\usage{
simulate_seir_cv(t, state_t0, param)
}
\arguments{
\item{t}{Number of time steps over which to sample from the model}

\item{state_t0}{Initial state of the model (see ?seir_state0)}

\item{param}{Model parameters (see ?seir_param)}
}
\value{
Object of class covoidd and dcm (from the package EpiModels)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

S = Susceptible
E = Exposed (latent period)
I = Infectious
R = Removed (recovered or died)
}
\examples{
# coming soon....

}
