% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-cm-model.R
\name{sir_c_param}
\alias{sir_c_param}
\title{SIR model parameters}
\usage{
sir_c_param(R0, gamma, cm)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> I), can be a function of t.}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}

\item{cm}{Contact matrix}
}
\value{
List of SIR model parameters
}
\description{
Setup function
}
