% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-cm-model.R
\name{sir_c_param}
\alias{sir_c_param}
\title{SIR model parameters}
\usage{
sir_c_param(R0, gamma, cm, dist, intervention = NULL)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> I), can be a function of t.}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}

\item{cm}{Contact matrix or a named list of contact matrices where the sum is taken to
be all contacts in the population.}

\item{dist}{Proportion of the population in each age category. Should sum to 1.}

\item{intervention}{An intervention created using `create_intervention` or named list
of interventions with the names matching the list of contact matrices.}
}
\value{
List of SIR model parameters
}
\description{
Setup function
}
