% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-cm-model.R
\name{sir_c_state0}
\alias{sir_c_state0}
\title{SIR model inital state}
\usage{
sir_c_state0(S0, I0, R0)
}
\arguments{
\item{S0}{Initial number of susceptibles}

\item{I0}{Initial number of infected}

\item{R0}{Initial number of removed}
}
\value{
List of SIR model initial states
}
\description{
Setup function
}
