% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sir-model.R
\name{sir_param}
\alias{sir_param}
\title{SIR model parameters}
\usage{
sir_param(R0, beta, gamma)
}
\arguments{
\item{R0}{Basic/empirical reproduction number (S -> I), can be a function of t.}

\item{beta}{Rate of potential new infections per infected (S -> I), can be a function of t.
Pass either R0 or beta.}

\item{gamma}{Inverse of the average length of infectious period (I -> R)}
}
\value{
List of SIR model parameters
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Setup function
}
